+++ source/db_constraints/constraints.d
@@ -9,42 +9,32 @@ import std.functional : unaryFun;
 KeyedItem will create a struct with *name* defined in the compile-time argument.
 For example a property marked with @UniqueColumn!("uc_Person") will
 be part of the struct uc_Person.
+Params:
+    name_ = The name of the constraint which is the structs name.
  */
 struct UniqueConstraintColumn(string name_)
 {
-/**
-The name of the constraint which is the structs name.
- */
     enum name = name_;
 }
 
-template helper(T...)
-{
-    alias T helper;
-}
-
 /**
-An alias for the primary key column.
+An alias for the primary key column. A member with this attribute
+must also have the NotNull attribute.
  */
-//alias PrimaryKeyColumn = helper!(UniqueConstraintColumn!("PrimaryKey"), NotNull);
 alias PrimaryKeyColumn = UniqueConstraintColumn!("PrimaryKey");
 
 
 /**
 KeyedItem.checkConstraints will check all of the members marked
 with this attribute and use the check given.
+Params:
+    check_ = The function that returns a boolean.
+    name_ = Name used in the error message if the function returns false.
  */
 struct CheckConstraint(alias check_, string name_ = "")
     if (is(typeof(unaryFun!check_)))
 {
-/**
-The function that returns a boolean.
- */
     alias check = unaryFun!check_;
-/**
-Name used in the error message if the function returns false. This may
-help narrow down which constraint failed.
- */
     enum name = name_;
 }
 
@@ -69,7 +59,6 @@ alias NotNull = CheckConstraint!(
             return true;
         }
     }, "NotNull");
-
 
 ///
 enum ForeignKeyActions
@@ -107,6 +96,13 @@ Updates or deletes the item based on what happened to the parent key.
 
 /**
 The foreign key user-defined attribute. Currently under :construction:
+Params:
+    childCols_ = The members in the child class that are used in the foreign key
+    parentTableName_ = The plural class name.
+    parentCols_ = The members in the parent class that are references in the foreign key
+    name_ = The name of the foreign key constraint. Will be used in error message when violated
+    onUpdate_ = What should happen when a foreign key is updated that is being referenced.
+    onDelete_ = What should happen when a foreign key is deleted that is being referenced.
  */
 struct ForeignKeyConstraint(string[] childCols_, string parentTableName_, string[] parentCols_,
                             string name_ = "",