+++ source/tests/one_to_one_class.d
@@ -25,7 +25,7 @@ unittest
     private:
         string _name;
     public:
-        @PrimaryKeyColumn
+        @PrimaryKeyColumn @NotNull
         string name() const @property nothrow pure @safe @nogc
         {
             return _name;